requirements.txt
# requirements.txt
langchain
langchain-google-genai  
chromadb
pypdf
python-dotenv
tiktoken
langchain-openai
langchain-community
langchain-chroma
pypdf
streamlit